pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentarena"
version = "0.1.0"
description = "A system that orchestrates contests between AI agents"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Bruce Kroeze", email = "bkroeze@gmail.com"},
]
dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.21.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
    "llm>=0.24.2",
    "sqlite-utils",
    "structlog>=25.3.0",
    "python-ulid[pydantic]",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.2.0",
    "flake8>=6.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["agentarena"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
